Test: Remove global getClientRects polyfill
