+++ packages/editor/src/components/collab-sidebar/rich-text-control/test/index.jsdom.test.tsx
@@ -10,7 +10,7 @@ import {
 	vi,
 	type Mock,
 } from 'vitest';
-import { useContext, useEffect, useState } from '@wordpress/element';
+import { useContext, useEffect } from '@wordpress/element';
 import { Fill, Popover, Slot, SlotFillProvider } from '@wordpress/components';
 import {
 	unregisterFormatType,
@@ -66,22 +66,6 @@ async function moveFocusTo( element: HTMLElement ) {
 	await flushMicrotasks();
 }
 
-// Type the way a browser would — `useRichText` builds its value by reading the
-// DOM and selection back from its own `input` handler.
-async function typeIntoTextbox( textbox: HTMLElement, text: string ) {
-	act( () => {
-		textbox.textContent = text;
-		const range = document.createRange();
-		range.selectNodeContents( textbox );
-		range.collapse( false );
-		const selection = document.getSelection();
-		selection?.removeAllRanges();
-		selection?.addRange( range );
-	} );
-	fireEvent.input( textbox, { inputType: 'insertText' } );
-	await flushMicrotasks();
-}
-
 describe( 'RichTextControl', () => {
 	beforeAll( () => {
 		// Register a minimal stub for `core/bold` so the optional
@@ -92,17 +76,10 @@ describe( 'RichTextControl', () => {
 			className: null,
 			edit: () => null,
 		} );
-
-		// The autocomplete popover measures the caret range to anchor itself,
-		// which jsdom leaves unimplemented. Any non-empty rect lets it mount.
-		Range.prototype.getClientRects = () =>
-			[ new DOMRect( 0, 0, 1, 1 ) ] as unknown as DOMRectList;
 	} );
 
 	afterAll( () => {
 		unregisterFormatType( 'core/bold' );
-		// @ts-expect-error -- Restore jsdom's own (absent) implementation.
-		delete Range.prototype.getClientRects;
 	} );
 
 	it( 'renders a labeled contenteditable textbox', () => {
@@ -160,48 +137,6 @@ describe( 'RichTextControl', () => {
 		);
 
 		expect( getTextbox( container ) ).toBeInTheDocument();
-	} );
-
-	it( 'points the textbox at the suggestions listbox once a completer matches', async () => {
-		// Held outside the completer: the options are reported back through an
-		// effect keyed on their identity, so rebuilding them per render loops.
-		const items = [ { key: 'alice', value: 'Alice', label: 'Alice' } ];
-		const completer = {
-			name: 'test/mentions',
-			triggerPrefix: '@',
-			useItems: () => [ items ],
-			getOptionCompletion: () => '@Alice',
-		};
-
-		function ControlledRichText() {
-			const [ value, setValue ] = useState( '' );
-			return (
-				<RichTextControl
-					label="Note"
-					value={ value }
-					onChange={ setValue }
-					completers={
-						[ completer ] as unknown as React.ComponentProps<
-							typeof RichTextControl
-						>[ 'completers' ]
-					}
-				/>
-			);
-		}
-
-		const { container } = render( <ControlledRichText /> );
-		const textbox = getTextbox( container );
-
-		await moveFocusTo( textbox );
-		await typeIntoTextbox( textbox, '@' );
-
-		const listbox = screen.getByRole( 'listbox' );
-		const option = screen.getByRole( 'option', { name: 'Alice' } );
-
-		expect( textbox ).toHaveAttribute( 'aria-autocomplete', 'list' );
-		expect( textbox ).toHaveAttribute( 'aria-haspopup', 'listbox' );
-		expect( textbox ).toHaveAttribute( 'aria-controls', listbox.id );
-		expect( textbox ).toHaveAttribute( 'aria-activedescendant', option.id );
 	} );
 
 	it( 'visually hides the label when `hideLabelFromVision` is set', () => {
