+++ test/unit/vitest-policy-exceptions.json
@@ -18,7 +18,6 @@
 		"packages/grid/src/dashboard-lanes/test/keyboard-activation.jsdom.test.tsx": "Stubs ResizeObserver only to mount the lanes; assertions cover the deterministic keyboard activator DOM hierarchy.",
 		"packages/grid/src/dashboard-lanes/test/use-lane-placement.jsdom.test.tsx": "Injects ResizeObserver entries and requestAnimationFrame scheduling to test deterministic lane-placement state without browser layout.",
 		"packages/ui/src/form/primitives/control-with-error/test/control-with-error.jsdom.test.tsx": "Uses requestAnimationFrame only to schedule reportValidity after React state; assertions cover aria-describedby semantics.",
-		"packages/editor/src/components/collab-sidebar/rich-text-control/test/index.jsdom.test.tsx": "Supplies a non-empty caret rectangle only so autocomplete can mount; assertions cover contenteditable labeling and text semantics.",
 		"packages/editor/src/components/collab-sidebar/test/board-store.jsdom.test.js": "Uses supplied DOMRect fixtures to test deterministic range-to-anchor calculations without relying on rendered layout."
 	}
 }
